refactor: Move type definitions to separate files
