+++ packages/vue/src/composables/use-global-cancel-stack/types.ts
@@ -1,0 +1,16 @@
+export interface UseGlobalCancelStackOptions {
+  /**
+   * @default
+   * ```ts
+   * true
+   * ```
+   */
+  escape?: boolean;
+  /**
+   * @default
+   * ```ts
+   * true
+   * ```
+   */
+  history?: boolean;
+}
