+++ packages/vue/src/composables/use-global-history-stack/index.ts
@@ -1,5 +1,8 @@
 import { findLastIndex } from "lodash-es";
 import type { MaybePromise } from "~/types";
+import type { UseGlobalHistoryStackRevokeOptions } from "./types";
+
+export * from "./types";
 
 export function useGlobalHistoryStack(handler: (event: PopStateEvent) => void) {
   function create() {
@@ -55,8 +58,4 @@ function handlePopState(event: PopStateEvent) {
   }
   if (!useGlobalHistoryStack.stack.length) removeWindowHandler();
   if (useGlobalHistoryStack.stack.length) history.pushState(null, "");
-}
-
-export interface UseGlobalHistoryStackRevokeOptions {
-  historyBack: boolean;
 }
