+++ packages/vue/src/composables/use-global-history-stack/types.ts
@@ -1,0 +1,3 @@
+export interface UseGlobalHistoryStackRevokeOptions {
+  historyBack: boolean;
+}
