+++ packages/vue/src/validate/composables/use-form/types.ts
@@ -1,0 +1,4 @@
+export type UseFormSubmitEventHandler = (event: UseFormSubmitEvent) => void;
+export type UseFormSubmitEvent =
+  | { type: "valid" }
+  | { type: "invalid"; order: number };
