+++ packages/vue/src/validate/utils/create-form/types.ts
@@ -1,0 +1,8 @@
+import type { FieldLike, FieldValidateResult } from "~/validate";
+
+export interface FormValidateResult {
+  valid: boolean;
+  fieldsResults: FieldValidateResult<string>[];
+  invalidFields: FieldLike[];
+  invalidIndexes: number[];
+}
